pyproject.toml
[build-system]
requires = ["setuptools>=75.1.0", "wheel>=0.45.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datashield"
version = "0.1.0"
description = "Data-Shield — Sensitive credential scanner and encryption vault for Windows 10/11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Data-Shield Team" }]
requires-python = ">=3.13"
keywords = ["security", "credentials", "scanner", "encryption", "vault", "windows"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Monitoring",
]

dependencies = [
    # GUI
    "PySide6>=6.8.2",
    # CLI & Display
    "rich>=14.0.0",
    "colorama>=0.4.6",
    "click>=8.1.8",
    # Detection engines
    "yara-python>=4.5.1",
    "regex>=2024.11.6",
    # Encryption
    "cryptography>=43.0.0",
    "bcrypt>=4.2.0",
    # Database & ORM
    "SQLAlchemy>=2.0.0",
    "alembic>=1.14.0",
    # Configuration
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    # File & System
    "watchdog>=5.0.0",
    "psutil>=6.1.0",
    "pywin32>=310.0",
    # Notifications
    "winotify>=1.1.0",
    # Utilities
    "platformdirs>=4.3.0",
    "humanize>=4.10.0",
    "loguru>=0.7.2",
    "attrs>=24.2.0",
    "python-dateutil>=2.8.2",
    # Scheduling
    "APScheduler>=3.11.0",
    # Structured logging
    "python-json-logger>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-qt>=4.4.0",
    "black>=24.4.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "bandit>=1.7.0",
]
build = [
    "PyInstaller>=6.10.0",
]

[project.scripts]
datashield = "datashield.__main__:main"
datashield-admin = "datashield.__main__:admin_main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
datashield = [
    "patterns/yara_rules/*.yar",
    "gui/resources/styles/*.qss",
    "gui/resources/icons/*.png",
    "gui/resources/icons/*.ico",
    "i18n/*.json",
    "config/defaults.toml",
]

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "W", "I", "N", "UP", "S", "B"]
ignore = ["S101", "S603", "S605", "S607"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/datashield --cov-report=term-missing --cov-report=html"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101", "B603", "B605", "B607"]
